Moved asyncio features into 'py3' folder due to async being a reserved word in Python 3.7 Deprecated mygeotab.async package
